paterns de comandes

+++ VR/Assets/Scripts/IEnnemy.cs
@@ -3,7 +3,7 @@
     public interface IEnnemy
     {
         void Walknig(float speed);
-        void Stop(int  timeToStop);
+        void Stop(float  timeToStop);
         void Attacking();
         void ReturnFromStart();
     }
